+++ src/main/java/roomescape/controller/ReservationController.java
@@ -2,6 +2,9 @@ package roomescape.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import roomescape.controller.dto.ReservationRequest;
+import roomescape.controller.dto.ReservationResponse;
+import roomescape.controller.dto.ReservationTimeResponse;
 import roomescape.domain.Reservation;
 import roomescape.service.ReservationService;
