+++ src/main/java/roomescape/controller/ThemeController.java
@@ -9,11 +9,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.controller.dto.ThemeRequest;
+import roomescape.controller.dto.ThemeResponse;
 import roomescape.domain.Theme;
 import roomescape.service.ThemeService;
 
 @RestController
-@RequestMapping("/theme")
+@RequestMapping("/themes")
 public class ThemeController {
 
     private final ThemeService themeService;
@@ -28,10 +30,8 @@ public class ThemeController {
     }
 
     @PostMapping
-    public ResponseEntity<ThemeResponse> createTime(
-            @RequestBody ThemeRequest themeRequest) {
-        Theme Theme = themeService.saveTime(themeRequest.name(), themeRequest.description(),
-                themeRequest.thumbnailUrl());
+    public ResponseEntity<ThemeResponse> createTime(@RequestBody ThemeRequest themeRequest) {
+        Theme Theme = themeService.saveTime(themeRequest.name(), themeRequest.description(), themeRequest.thumbnailUrl());
         return ResponseEntity.ok(ThemeResponse.from(Theme));
     }
