--- src/main/java/roomescape/controller/ThemeResponse.java
@@ -1,8 +1,0 @@
-package roomescape.controller;
-
-import roomescape.domain.Theme;
-
-public record ThemeResponse() {
-    public static ThemeResponse from(Theme theme) {
-    }
-}
