+++ src/main/java/roomescape/controller/dto/ThemeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.dto;
 
 public record ThemeRequest(String name, String description, String thumbnailUrl) {
 }
